requirements.txt
alembic==0.9.9
aniso8601==9.0.1
appdirs==1.4.4
asgiref==3.3.4
beautifulsoup4==4.9.3
blinker==1.4
bs4==0.0.1
certifi==2020.12.5
cfgv==3.3.0
chardet==3.0.4
click==6.7
cssselect==1.1.0
distlib==0.3.1
fake-useragent==0.1.11
fastapi==0.65.2
filelock==3.0.12
Flask==1.0.2
Flask-Dance==0.14.0
Flask-DebugToolbar==0.10.1
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-OAuth==0.12
Flask-OAuthlib==0.9.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
greenlet==1.1.0
h11==0.12.0
httplib2==0.11.3
identify==2.2.4
idna==2.6
itsdangerous==0.24
Jinja2==2.10
lazy==1.3
loguru==0.5.3
lxml==4.6.3
Mako==1.0.7
Markdown==3.3.4
MarkupSafe==1.1.1
nodeenv==1.6.0
oauth2==1.9.0.post1
oauthlib==2.0.7
packaging==20.9
parse==1.19.0
Pillow==8.2.0
pluggy==0.13.1
pre-commit==2.12.1
py==1.10.0
pydantic==1.8.2
pyee==8.1.0
PyExecJS==1.5.1
PyJWT==1.4.2
pyparsing==2.4.7
PyPDF2==1.26.0
pyppeteer==0.2.5
pyquery==1.4.3
PySocks==1.7.1
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2021.1
PyYAML==5.4.1
requests==2.18.4
requests-html==0.10.0
requests-oauthlib==0.8.0
selenium==3.141.0
simplekv==0.14.1
six==1.11.0
soupsieve==2.2.1
SQLAlchemy==1.2.6
SQLAlchemy-Utils==0.33.2
starlette==0.14.2
toml==0.10.2
tox==3.23.1
tqdm==4.61.0
translate==3.5.0
tweepy==3.10.0
typing-extensions==3.10.0.0
urllib3==1.22
URLObject==2.4.3
uvicorn==0.14.0
virtualenv==20.4.6
w3lib==1.22.0
websockets==8.1
Werkzeug==0.14.1
wincertstore==0.2
WTForms==2.1
